requirements.txt
# Testing
mock==2.0.0
django-nose==1.4.4
spec==1.4.0
coverage==4.4.1
django-dynamic-fixture
tox==2.7.0
django-localflavor==1.0
